refactor: move esp32 program function from Brain to TestBed
also move targetResetPin to TestBed
